refactor(server/find-similar): 调整图模型，引入覆盖关系，调整存储结构、API结构和对应的实现

+++ server/src/main/kotlin/com/heerkirov/hedge/server/functions/service/FindSimilarService.kt
@@ -11,21 +11,21 @@ import com.heerkirov.hedge.server.dto.form.FindSimilarResultResolveForm
 import com.heerkirov.hedge.server.dto.form.FindSimilarTaskCreateForm
 import com.heerkirov.hedge.server.dto.res.*
 import com.heerkirov.hedge.server.events.SimilarFinderResultDeleted
-import com.heerkirov.hedge.server.events.SimilarFinderResultResolved
+import com.heerkirov.hedge.server.events.SimilarFinderResultUpdated
 import com.heerkirov.hedge.server.exceptions.*
 import com.heerkirov.hedge.server.functions.manager.BookManager
 import com.heerkirov.hedge.server.functions.manager.IllustManager
+import com.heerkirov.hedge.server.model.FindSimilarIgnored
 import com.heerkirov.hedge.server.model.Illust
-import com.heerkirov.hedge.server.utils.Json.parseJSONObject
-import com.heerkirov.hedge.server.utils.Json.toJsonNode
 import com.heerkirov.hedge.server.utils.business.filePathOrNullFrom
 import com.heerkirov.hedge.server.utils.business.map
 import com.heerkirov.hedge.server.utils.business.toListResult
 import com.heerkirov.hedge.server.utils.ktorm.OrderTranslator
+import com.heerkirov.hedge.server.utils.ktorm.firstOrNull
 import com.heerkirov.hedge.server.utils.ktorm.orderBy
+import com.heerkirov.hedge.server.utils.tuples.Tuple6
 import com.heerkirov.hedge.server.utils.types.descendingOrderItem
 import org.ktorm.dsl.*
-import org.ktorm.entity.Tuple4
 import org.ktorm.entity.firstOrNull
 import org.ktorm.entity.sequenceOf
 import java.time.Instant
@@ -73,152 +73,176 @@ class FindSimilarService(private val data: DataRepository,
 
     fun listResult(filter: LimitAndOffsetFilter): ListResult<FindSimilarResultRes> {
         val results = data.db.from(FindSimilarResults)
-            .select(FindSimilarResults.id, FindSimilarResults.summaryTypes, FindSimilarResults.images, FindSimilarResults.recordTime)
+            .select(FindSimilarResults.id, FindSimilarResults.category, FindSimilarResults.summaryType, FindSimilarResults.imageIds, FindSimilarResults.resolved, FindSimilarResults.recordTime)
             .limit(filter.offset, filter.limit)
-            .orderBy(FindSimilarResults.sortPriority.desc(), FindSimilarResults.recordTime.desc())
+            .orderBy(FindSimilarResults.category.asc(), FindSimilarResults.recordTime.desc())
             .toListResult {
                 val id = it[FindSimilarResults.id]!!
-                val summaryTypes = it[FindSimilarResults.summaryTypes]!!
-                val images = it[FindSimilarResults.images]!!.map { i -> i.toInt() }
+                val category = it[FindSimilarResults.category]!!
+                val summaryType = it[FindSimilarResults.summaryType]!!
+                val imageIds = it[FindSimilarResults.imageIds]!!
+                val resolved = it[FindSimilarResults.resolved]!!
                 val recordTime = it[FindSimilarResults.recordTime]!!
-                Tuple4(id, summaryTypes, images, recordTime)
+                Tuple6(id, category, summaryType, imageIds, resolved, recordTime)
             }
 
-        val illustIds = results.result.asSequence().map { (_, _, images, _) -> images }.flatten().toList()
+        val illustIds = results.result.asSequence().map { (_, _, _, images, _, _) -> images }.flatten().toList()
 
         val imageFiles = data.db.from(Illusts)
             .innerJoin(FileRecords, FileRecords.id eq Illusts.fileId)
             .select(Illusts.id, FileRecords.id, FileRecords.block, FileRecords.extension, FileRecords.status)
             .where { Illusts.id inList illustIds }
-            .associate { it[Illusts.id]!! to filePathOrNullFrom(it) }
+            .associateBy({ it[Illusts.id]!! }) { filePathOrNullFrom(it) }
 
-        return results.map { (id, s, i, r) ->
-            val images = i.map {
-                val filePath = imageFiles[it]
-                FindSimilarResultImage(it, filePath)
-            }
-            FindSimilarResultRes(id, s, images, r)
+        return results.map { (id, c, s, i, d, r) ->
+            val images = i.map { FindSimilarResultImage(it, imageFiles[it]) }
+            FindSimilarResultRes(id, c, s, images, d, r)
         }
     }
 
     fun getResult(id: Int): FindSimilarResultDetailRes {
         val result = data.db.sequenceOf(FindSimilarResults).firstOrNull { it.id eq id } ?: throw be(NotFound())
 
-        val illustIds = result.images.map { it.toInt() }.toList()
-
         val imageFiles = data.db.from(Illusts)
             .innerJoin(FileRecords, FileRecords.id eq Illusts.fileId)
             .select(Illusts.id, FileRecords.id, FileRecords.block, FileRecords.extension, FileRecords.status)
-            .where { Illusts.id inList illustIds }
+            .where { Illusts.id inList result.imageIds }
             .associate { it[Illusts.id]!! to filePathOrNullFrom(it) }
 
-        val images = illustIds.map {
-            val filePath = imageFiles[it]
-            FindSimilarResultImage(it, filePath)
-        }
+        val images = result.imageIds.map { FindSimilarResultImage(it, imageFiles[it]) }
 
-        val relations = result.relations.map { FindSimilarResultRelation(it.a, it.b, it.type, it.params) }
-
-        return FindSimilarResultDetailRes(result.id, result.summaryTypes, images, relations, result.recordTime)
+        return FindSimilarResultDetailRes(result.id, result.category, result.summaryType, images, result.edges, result.coverages, result.resolved, result.recordTime)
     }
 
     /**
-     * @throws ResourceNotExist ("config.a"|"config.b", EntityKey)
+     * @throws ResourceNotExist ("from"|"to", number)
+     * @throws ResourceNotExist ("imageIds", number[])
      */
     fun resolveResult(id: Int, form: FindSimilarResultResolveForm) {
         data.db.transaction {
             val result = data.db.sequenceOf(FindSimilarResults).firstOrNull { it.id eq id } ?: throw be(NotFound())
 
-            val entityKeys = result.images.map { it.toInt() }.toSet()
+            val entityKeys = result.imageIds.toSet()
 
             //actions参数校验
-            val actions = form.actions.map { action ->
-                when(action.actionType) {
-                    FindSimilarResultResolveForm.ActionType.CLONE_IMAGE,
-                    FindSimilarResultResolveForm.ActionType.MARK_IGNORED -> {
-                        if(action.b == null) throw be(ParamRequired("config.b"))
+            for (action in form.actions) {
+                when(action) {
+                    is FindSimilarResultResolveForm.ResolutionForTwoImage -> if(action.from !in entityKeys) {
+                        throw be(ResourceNotExist("from", action.from))
+                    }else if(action.to !in entityKeys) {
+                        throw be(ResourceNotExist("to", action.to))
                     }
-                    FindSimilarResultResolveForm.ActionType.DELETE,
-                    FindSimilarResultResolveForm.ActionType.ADD_TO_BOOK,
-                    FindSimilarResultResolveForm.ActionType.ADD_TO_COLLECTION -> {
-                        if(action.b != null) throw be(ParamNotRequired("config.b"))
+                    is FindSimilarResultResolveForm.ResolutionForMultipleImage -> if(!entityKeys.containsAll(action.imageIds)) {
+                        throw be(ResourceNotExist("imageIds", action.imageIds.filter { it !in entityKeys }))
                     }
+                    else -> {}
                 }
-
-                if(action.a !in entityKeys) throw be(ResourceNotExist("config.a", action.a))
-                if(action.b != null && action.b !in entityKeys) throw be(ResourceNotExist("config.b", action.b))
-
-                val config: Any? = try {
-                    when(action.actionType) {
-                        FindSimilarResultResolveForm.ActionType.CLONE_IMAGE -> action.config?.toJsonNode()?.parseJSONObject<FindSimilarResultResolveForm.CloneImageConfig>() ?: throw be(ParamRequired("config"))
-                        FindSimilarResultResolveForm.ActionType.ADD_TO_COLLECTION -> action.config?.toJsonNode()?.parseJSONObject<FindSimilarResultResolveForm.AddToCollectionConfig>() ?: throw be(ParamRequired("config"))
-                        FindSimilarResultResolveForm.ActionType.ADD_TO_BOOK -> action.config?.toJsonNode()?.parseJSONObject<FindSimilarResultResolveForm.AddToBookConfig>() ?: throw be(ParamRequired("config"))
-                        else -> null
-                    }
-                }catch (e: Exception) {
-                    throw be(ParamError("config"))
-                }
-
-                FindSimilarResultResolveForm.Resolution(action.a, action.b, action.actionType, config)
             }
 
             //actions依次执行。其中对image的collection/book更新会被收集起来批量执行
             val collectionToImages = mutableMapOf<Any, MutableList<Int>>()
             val bookToImages = mutableMapOf<Int, MutableList<Int>>()
-            for (action in actions) {
-                when(action.actionType) {
-                    FindSimilarResultResolveForm.ActionType.CLONE_IMAGE -> {
-                        val config = action.config as FindSimilarResultResolveForm.CloneImageConfig
-                        illustManager.cloneProps(action.a, action.b!!, config.props, config.merge, config.deleteFrom)
+            val toBeDeleted = mutableSetOf<Int>()
+            for (action in form.actions) {
+                when(action) {
+                    is FindSimilarResultResolveForm.CloneImageResolution -> {
+                        illustManager.cloneProps(action.from, action.to, action.props, action.merge, action.deleteFrom)
                     }
-                    FindSimilarResultResolveForm.ActionType.ADD_TO_COLLECTION -> {
-                        val config = action.config as FindSimilarResultResolveForm.AddToCollectionConfig
-                        collectionToImages.computeIfAbsent(config.collectionId) { mutableListOf() }.add(action.a)
+                    is FindSimilarResultResolveForm.AddToCollectionResolution -> {
+                        collectionToImages.computeIfAbsent(action.collectionId) { mutableListOf() }.addAll(action.imageIds)
                     }
-                    FindSimilarResultResolveForm.ActionType.ADD_TO_BOOK -> {
-                        val config = action.config as FindSimilarResultResolveForm.AddToBookConfig
-                        bookToImages.computeIfAbsent(config.bookId) { mutableListOf() }.add(action.a)
+                    is FindSimilarResultResolveForm.AddToBookResolution -> {
+                        bookToImages.computeIfAbsent(action.bookId) { mutableListOf() }.addAll(action.imageIds)
                     }
-                    FindSimilarResultResolveForm.ActionType.DELETE -> {
-                        val illust = data.db.sequenceOf(Illusts).firstOrNull { it.id eq action.a }
-                        if(illust != null) illustManager.delete(illust)
+                    is FindSimilarResultResolveForm.DeleteResolution -> {
+                        toBeDeleted.addAll(action.imageIds)
                     }
-                    FindSimilarResultResolveForm.ActionType.MARK_IGNORED -> {
-                        val ak = action.a.toString()
-                        val bk = action.b!!.toString()
+                    is FindSimilarResultResolveForm.MarkIgnoredResolution -> {
                         val exist = data.db.from(FindSimilarIgnores)
                             .select((count() greater 0).aliased("exist"))
-                            .where { (FindSimilarIgnores.firstTarget eq ak) and (FindSimilarIgnores.secondTarget eq bk) }
+                            .where { (FindSimilarIgnores.type eq FindSimilarIgnored.IgnoredType.EDGE) and (FindSimilarIgnores.firstTarget eq action.from) and (FindSimilarIgnores.secondTarget eq action.to) }
                             .map { it.getBoolean("exist") }
                             .first()
 
                         if(!exist) {
                             val now = Instant.now()
                             data.db.insert(FindSimilarIgnores) {
-                                set(it.firstTarget, ak)
-                                set(it.secondTarget, bk)
+                                set(it.type, FindSimilarIgnored.IgnoredType.EDGE)
+                                set(it.firstTarget, action.from)
+                                set(it.secondTarget, action.to)
                                 set(it.recordTime, now)
                             }
                             data.db.insert(FindSimilarIgnores) {
-                                set(it.firstTarget, bk)
-                                set(it.secondTarget, ak)
+                                set(it.type, FindSimilarIgnored.IgnoredType.EDGE)
+                                set(it.firstTarget, action.to)
+                                set(it.secondTarget, action.from)
                                 set(it.recordTime, now)
+                            }
+                        }
+                    }
+                    is FindSimilarResultResolveForm.MarkIgnoredSourceBookResolution -> {
+                        val sourceBookId = data.db.from(SourceBooks)
+                            .select(SourceBooks.id)
+                            .where { SourceBooks.site eq action.site and (SourceBooks.code eq action.sourceBookCode) }
+                            .firstOrNull()
+                            ?.get(SourceBooks.id)
+
+                        if(sourceBookId != null) {
+                            val exist = data.db.from(FindSimilarIgnores)
+                                .select((count() greater 0).aliased("exist"))
+                                .where { (FindSimilarIgnores.type eq FindSimilarIgnored.IgnoredType.SOURCE_BOOK) and (FindSimilarIgnores.firstTarget eq sourceBookId) }
+                                .map { it.getBoolean("exist") }
+                                .first()
+
+                            if(!exist) {
+                                val now = Instant.now()
+                                data.db.insert(FindSimilarIgnores) {
+                                    set(it.type, FindSimilarIgnored.IgnoredType.SOURCE_BOOK)
+                                    set(it.firstTarget, sourceBookId)
+                                    set(it.secondTarget, null)
+                                    set(it.recordTime, now)
+                                }
+                            }
+                        }
+
+                    }
+                    is FindSimilarResultResolveForm.MarkIgnoredSourceDataResolution -> {
+                        val sourceDataId = data.db.from(SourceDatas)
+                            .select(SourceDatas.id)
+                            .where { SourceDatas.sourceSite eq action.site and (SourceDatas.sourceId eq action.sourceId) }
+                            .firstOrNull()
+                            ?.get(SourceDatas.id)
+
+                        if(sourceDataId != null) {
+                            val exist = data.db.from(FindSimilarIgnores)
+                                .select((count() greater 0).aliased("exist"))
+                                .where { (FindSimilarIgnores.type eq FindSimilarIgnored.IgnoredType.SOURCE_IDENTITY_SIMILAR) and (FindSimilarIgnores.firstTarget eq sourceDataId) }
+                                .map { it.getBoolean("exist") }
+                                .first()
+
+                            if(!exist) {
+                                val now = Instant.now()
+                                data.db.insert(FindSimilarIgnores) {
+                                    set(it.type, FindSimilarIgnored.IgnoredType.SOURCE_IDENTITY_SIMILAR)
+                                    set(it.firstTarget, sourceDataId)
+                                    set(it.secondTarget, null)
+                                    set(it.recordTime, now)
+                                }
                             }
                         }
                     }
                 }
             }
 
-            //有关collection有些特别机制，因此需要全部提取到一起执行。
-            //collectionId可以设置为string，表示导入后会创建一个新collection。
-            for ((collectionId, entities) in collectionToImages) {
+            for ((collectionId, imageIds) in collectionToImages) {
                 when (collectionId) {
                     is Int -> {
-                        val images = illustManager.unfoldImages(entities + listOf(collectionId), sorted = false)
+                        val images = illustManager.unfoldImages(imageIds + listOf(collectionId), sorted = false)
                         illustManager.updateImagesInCollection(collectionId, images)
                     }
-                    is String -> if(entities.isNotEmpty()) {
-                        illustManager.newCollection(entities, "", null, null, Illust.Tagme.EMPTY)
+                    is String -> if(imageIds.isNotEmpty()) {
+                        //collectionId可以设置为string，表示会创建新collection，相同字符串的会被创建到同一个collection中
+                        illustManager.newCollection(imageIds, "", null, null, Illust.Tagme.EMPTY)
                     }
                     else -> throw be(ParamTypeError("config.collectionId", "must be number or string."))
                 }
@@ -226,10 +250,17 @@ class FindSimilarService(private val data: DataRepository,
             for ((bookId, imageIds) in bookToImages) {
                 bookManager.addImagesInBook(bookId, imageIds, null)
             }
+            if (toBeDeleted.isNotEmpty()) {
+                illustManager.unfoldImages(toBeDeleted.toList()).forEach(illustManager::delete)
+            }
 
-            data.db.delete(FindSimilarResults) { it.id eq id }
+            if (form.clear) {
+                data.db.delete(FindSimilarResults) { it.id eq id }
+                bus.emit(SimilarFinderResultDeleted(id))
+            }else{
+                bus.emit(SimilarFinderResultUpdated(id))
+            }
 
-            bus.emit(SimilarFinderResultResolved(id))
         }
     }
